+++ packages/ragbits-evaluate/src/ragbits/evaluate/__init__.py
@@ -1,4 +1,4 @@
 from .evaluator import BaseEvaluator, Evaluator
 from .optimizer import BaseOptimizer, Optimizer
 
-__all__ = ["Optimizer", "Evaluator", "BaseOptimizer", "BaseEvaluator"]
+__all__ = ["BaseEvaluator", "BaseOptimizer", "Evaluator", "Optimizer"]